Add can now print transition table to file, test for ToFile, new file Tablegenerator to reduce code duplication
